+++ pkgs/test_core/lib/src/runner/configuration.dart
@@ -58,11 +58,22 @@ class Configuration {
   final bool? _pauseAfterLoad;
 
   /// Whether to run browsers in their respective debug modes
-  bool get debug => pauseAfterLoad || (_debug ?? false) || coverage != null;
+  bool get debug =>
+      pauseAfterLoad ||
+      (_debug ?? false) ||
+      coverage != null ||
+      coverageLcov != null;
   final bool? _debug;
 
-  /// The output folder for coverage gathering
+  /// The output folder for coverage gathering.
   final String? coverage;
+
+  /// The lcov file to output coverage to.
+  final String? coverageLcov;
+
+  /// Whether to collect branch coverage info.
+  bool get branchCoverage => _branchCoverage ?? false;
+  final bool? _branchCoverage;
 
   /// The path to the file from which to load more configuration information.
   ///
@@ -260,6 +271,8 @@ class Configuration {
     required String? reporter,
     required Map<String, String>? fileReporters,
     required String? coverage,
+    required String? coverageLcov,
+    required bool? branchCoverage,
     required int? concurrency,
     required int? shardIndex,
     required int? totalShards,
@@ -313,6 +326,8 @@ class Configuration {
       reporter: reporter,
       fileReporters: fileReporters,
       coverage: coverage,
+      coverageLcov: coverageLcov,
+      branchCoverage: branchCoverage,
       concurrency: concurrency,
       shardIndex: shardIndex,
       totalShards: totalShards,
@@ -372,6 +387,8 @@ class Configuration {
     String? reporter,
     Map<String, String>? fileReporters,
     String? coverage,
+    String? coverageLcov,
+    bool? branchCoverage,
     int? concurrency,
     int? shardIndex,
     int? totalShards,
@@ -423,6 +440,8 @@ class Configuration {
     reporter: reporter,
     fileReporters: fileReporters,
     coverage: coverage,
+    coverageLcov: coverageLcov,
+    branchCoverage: branchCoverage,
     concurrency: concurrency,
     shardIndex: shardIndex,
     totalShards: totalShards,
@@ -491,6 +510,8 @@ class Configuration {
     reporter: null,
     fileReporters: null,
     coverage: null,
+    coverageLcov: null,
+    branchCoverage: null,
     concurrency: null,
     shardIndex: null,
     totalShards: null,
@@ -555,6 +576,8 @@ class Configuration {
     reporter: null,
     fileReporters: null,
     coverage: null,
+    coverageLcov: null,
+    branchCoverage: null,
     concurrency: null,
     shardIndex: null,
     totalShards: null,
@@ -622,6 +645,8 @@ class Configuration {
     color: null,
     configurationPath: null,
     coverage: null,
+    coverageLcov: null,
+    branchCoverage: null,
     shardIndex: null,
     totalShards: null,
     testSelections: null,
@@ -683,6 +708,8 @@ class Configuration {
     reporter: null,
     fileReporters: null,
     coverage: null,
+    coverageLcov: null,
+    branchCoverage: null,
     concurrency: null,
     shardIndex: null,
     totalShards: null,
@@ -750,6 +777,8 @@ class Configuration {
     required String? reporter,
     required Map<String, String>? fileReporters,
     required this.coverage,
+    required this.coverageLcov,
+    required bool? branchCoverage,
     required int? concurrency,
     required this.shardIndex,
     required this.totalShards,
@@ -776,6 +805,7 @@ class Configuration {
        _configurationPath = configurationPath,
        _reporter = reporter,
        fileReporters = fileReporters ?? {},
+       _branchCoverage = branchCoverage,
        _concurrency = concurrency,
        _testSelections =
            testSelections == null || testSelections.isEmpty
@@ -841,6 +871,8 @@ class Configuration {
     reporter: null,
     fileReporters: null,
     coverage: null,
+    coverageLcov: null,
+    branchCoverage: null,
     concurrency: null,
     shardIndex: null,
     totalShards: null,
@@ -941,6 +973,8 @@ class Configuration {
       reporter: other._reporter ?? _reporter,
       fileReporters: mergeMaps(fileReporters, other.fileReporters),
       coverage: other.coverage ?? coverage,
+      coverageLcov: other.coverageLcov ?? coverageLcov,
+      branchCoverage: other._branchCoverage ?? _branchCoverage,
       concurrency: other._concurrency ?? _concurrency,
       shardIndex: other.shardIndex ?? shardIndex,
       totalShards: other.totalShards ?? totalShards,
@@ -996,6 +1030,8 @@ class Configuration {
     String? reporter,
     Map<String, String>? fileReporters,
     String? coverage,
+    String? coverageLcov,
+    bool? branchCoverage,
     int? concurrency,
     int? shardIndex,
     int? totalShards,
@@ -1044,6 +1080,8 @@ class Configuration {
       reporter: reporter ?? _reporter,
       fileReporters: fileReporters ?? this.fileReporters,
       coverage: coverage ?? this.coverage,
+      coverageLcov: coverageLcov ?? this.coverageLcov,
+      branchCoverage: branchCoverage ?? _branchCoverage,
       concurrency: concurrency ?? _concurrency,
       shardIndex: shardIndex ?? this.shardIndex,
       totalShards: totalShards ?? this.totalShards,