+++ pkgs/test_core/lib/src/runner/coverage_stub.dart
@@ -2,11 +2,21 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:coverage/coverage.dart';
+
 import 'live_suite_controller.dart';
 
-Future<void> writeCoverage(
-  String coveragePath,
+Future<Map<String, HitMap>> writeCoverage(
+  String? coveragePath,
   LiveSuiteController controller,
+) =>
+    throw UnsupportedError(
+      'Coverage is only supported through the test runner.',
+    );
+
+Future<void> writeCoverageLcov(
+  String coverageLcov,
+  Map<String, HitMap> allCoverageData,
 ) =>
     throw UnsupportedError(
       'Coverage is only supported through the test runner.',